+++ Source/PrismLibrary_XF.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.14
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Prism", "Prism\Prism.csproj", "{E6C50355-D01E-4CAA-884D-D7929861315C}"
 EndProject
@@ -72,6 +72,7 @@ Global
 		{79EBB8F1-5D80-4EA1-9100-EBDC1EC51CB7}.Release|x64.ActiveCfg = Release|Any CPU
 		{79EBB8F1-5D80-4EA1-9100-EBDC1EC51CB7}.Release|x86.ActiveCfg = Release|Any CPU
 		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Debug|ARM.ActiveCfg = Debug|Any CPU
 		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -81,6 +82,7 @@ Global
 		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Release|x64.ActiveCfg = Release|Any CPU
 		{B6E74918-D92B-4690-BEC9-0BB199A6E933}.Release|x86.ActiveCfg = Release|Any CPU
 		{DD171BA8-7B2D-4722-B5BF-E0A9EBFA1D5F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DD171BA8-7B2D-4722-B5BF-E0A9EBFA1D5F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DD171BA8-7B2D-4722-B5BF-E0A9EBFA1D5F}.Debug|ARM.ActiveCfg = Debug|Any CPU
 		{DD171BA8-7B2D-4722-B5BF-E0A9EBFA1D5F}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{DD171BA8-7B2D-4722-B5BF-E0A9EBFA1D5F}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -110,6 +112,7 @@ Global
 		{174F4ADD-B9C7-4F5E-A96D-F53FF3ECBE9B}.Release|x64.ActiveCfg = Release|Any CPU
 		{174F4ADD-B9C7-4F5E-A96D-F53FF3ECBE9B}.Release|x86.ActiveCfg = Release|Any CPU
 		{1D1B893C-3F38-436F-A507-D65310B9D2A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D1B893C-3F38-436F-A507-D65310B9D2A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D1B893C-3F38-436F-A507-D65310B9D2A0}.Debug|ARM.ActiveCfg = Debug|Any CPU
 		{1D1B893C-3F38-436F-A507-D65310B9D2A0}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{1D1B893C-3F38-436F-A507-D65310B9D2A0}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -129,6 +132,7 @@ Global
 		{943664CF-3A0C-4EFF-8BDA-29241D49DD85}.Release|x64.ActiveCfg = Release|Any CPU
 		{943664CF-3A0C-4EFF-8BDA-29241D49DD85}.Release|x86.ActiveCfg = Release|Any CPU
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Debug|ARM.ActiveCfg = Debug|Any CPU
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -138,6 +142,7 @@ Global
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Release|x64.ActiveCfg = Release|Any CPU
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29}.Release|x86.ActiveCfg = Release|Any CPU
 		{2B5E37A2-B719-478A-A513-FBFA08F7EED6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B5E37A2-B719-478A-A513-FBFA08F7EED6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B5E37A2-B719-478A-A513-FBFA08F7EED6}.Debug|ARM.ActiveCfg = Debug|Any CPU
 		{2B5E37A2-B719-478A-A513-FBFA08F7EED6}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{2B5E37A2-B719-478A-A513-FBFA08F7EED6}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -157,5 +162,8 @@ Global
 		{1D1B893C-3F38-436F-A507-D65310B9D2A0} = {FEC22018-CBB0-493A-942D-59D1B93496E8}
 		{0CB2C41D-2620-4FAF-9F73-FEAE52D93F29} = {FEC22018-CBB0-493A-942D-59D1B93496E8}
 		{2B5E37A2-B719-478A-A513-FBFA08F7EED6} = {FEC22018-CBB0-493A-942D-59D1B93496E8}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D55096A8-03B8-43A8-8372-DC93F8F093FD}
 	EndGlobalSection
 EndGlobal
