+++ Source/Xamarin/Prism.Forms.Tests/Mocks/Views/CarouselPageMock.cs
@@ -21,10 +21,7 @@ namespace Prism.Forms.Tests.Mocks.Views
             Children.Add(new ContentPage() { Title = "Page 3" });
 
             PageNavigationEventRecorder = recorder;
-
-            var recordable = BindingContext as IPageNavigationEventRecordable;
-            if (recordable != null)
-                recordable.PageNavigationEventRecorder = recorder;
+            ((IPageNavigationEventRecordable)BindingContext).PageNavigationEventRecorder = recorder;
         }
 
         public void Destroy()
