+++ Source/Xamarin/Prism.Forms.Tests/Mocks/Views/NavigationPageMock.cs
@@ -23,10 +23,7 @@ namespace Prism.Forms.Tests.Mocks.Views
             ViewModelLocator.SetAutowireViewModel(this, true);
 
             PageNavigationEventRecorder = recorder;
-
-            var recordable = BindingContext as IPageNavigationEventRecordable;
-            if (recordable != null)
-                recordable.PageNavigationEventRecorder = recorder;
+            ((IPageNavigationEventRecordable)BindingContext).PageNavigationEventRecorder = recorder;
         }
 
         public void Destroy()
@@ -65,10 +62,7 @@ namespace Prism.Forms.Tests.Mocks.Views
             ViewModelLocator.SetAutowireViewModel(this, true);
 
             PageNavigationEventRecorder = recorder;
-
-            var recordable = BindingContext as IPageNavigationEventRecordable;
-            if (recordable != null)
-                recordable.PageNavigationEventRecorder = recorder;
+            ((IPageNavigationEventRecordable)BindingContext).PageNavigationEventRecorder = recorder;
         }
 
         public void OnNavigatedFrom(NavigationParameters parameters)
