+++ Source/Xamarin/Prism.Forms.Tests/Mocks/Views/TabbedPageMock.cs
@@ -18,7 +18,7 @@ namespace Prism.Forms.Tests.Mocks.Views
             ViewModelLocator.SetAutowireViewModel(this, true);
 
             Children.Add(new ContentPageMock(recorder) { Title = "Page 1" });
-            Children.Add(new PageMock() { Title = "Page 2" });
+            Children.Add(new PageMock() { Title = "Page 2", BindingContext = null });
             Children.Add(new ContentPageMock(recorder) { Title = "Page 3" });
 
             PageNavigationEventRecorder = recorder;
