+++ Source/Xamarin/Prism.Forms.Tests/Mvvm/Mocks/ViewModels/ViewModelLocatorPageMockViewModel.cs
@@ -1,0 +1,8 @@
+﻿using Prism.Forms.Tests.Mocks.ViewModels;
+
+namespace Prism.Forms.Tests.Mvvm.Mocks.ViewModels
+{
+    public class ViewModelLocatorPageMockViewModel : ViewModelBase
+    {
+    }
+}
