+++ Source/Xamarin/Prism.Forms.Tests/Mvvm/Mocks/Views/ViewModelLocatorPageMock.cs
@@ -1,0 +1,8 @@
+﻿using Xamarin.Forms;
+
+namespace Prism.Forms.Tests.Mvvm.Mocks.Views
+{
+    public class ViewModelLocatorPageMock : Page
+    {
+    }
+}
