+++ Source/Xamarin/Prism.Forms.Tests/Mvvm/ViewModelLocatorFixture.cs
@@ -1,13 +1,8 @@
-﻿using Prism.Forms.Tests.Mocks;
-using Prism.Forms.Tests.Mocks.ViewModels;
-using Prism.Forms.Tests.Mocks.Views;
+﻿using Prism.Forms.Tests.Mvvm.Mocks.ViewModels;
+using Prism.Forms.Tests.Mvvm.Mocks.Views;
 using Prism.Mvvm;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 using Xamarin.Forms;
 using Xunit;
 
@@ -20,13 +15,15 @@ namespace Prism.Forms.Tests.Mvvm
         {
             ResetViewModelLocationProvider();
 
-            PageMock view = new PageMock();
+            ViewModelLocatorPageMock view = new ViewModelLocatorPageMock();
             Assert.Null(view.BindingContext);
 
             ViewModelLocator.SetAutowireViewModel(view, true);
 
             Assert.NotNull(view.BindingContext);
-            Assert.IsType< PageMockViewModel>(view.BindingContext);
+            Assert.IsType<ViewModelLocatorPageMockViewModel>(view.BindingContext);
+
+            ResetViewModelLocationProvider();
         }
 
         [Fact]
@@ -34,11 +31,13 @@ namespace Prism.Forms.Tests.Mvvm
         {
             ResetViewModelLocationProvider();
 
-            PageMock view = new PageMock();
+            ViewModelLocatorPageMock view = new ViewModelLocatorPageMock();
 
             ViewModelLocator.SetAutowireViewModel(view, true);
 
             Assert.True(ViewModelLocator.GetAutowireViewModel(view));
+
+            ResetViewModelLocationProvider();
         }
 
         [Fact]
@@ -46,7 +45,7 @@ namespace Prism.Forms.Tests.Mvvm
         {
             ResetViewModelLocationProvider();
 
-            PageMock view = new PageMock();
+            ViewModelLocatorPageMock view = new ViewModelLocatorPageMock();
             Assert.Null(view.BindingContext);
 
             object mockObject = new object();
@@ -55,6 +54,8 @@ namespace Prism.Forms.Tests.Mvvm
             ViewModelLocator.SetAutowireViewModel(view, true);
             Assert.NotNull(view.BindingContext);
             Assert.Equal(mockObject, view.BindingContext);
+
+            ResetViewModelLocationProvider();
         }
 
         [Fact]
@@ -77,6 +78,8 @@ namespace Prism.Forms.Tests.Mvvm
             ViewModelLocator.SetAutowireViewModel(view, true);
             Assert.NotNull(view.BindingContext);
             Assert.IsType<ViewModelLocatorFixture>(view.BindingContext);
+
+            ResetViewModelLocationProvider();
         }
 
         private class TestPageOfShouldUseCustomDefaultViewTypeToViewModelTypeResolverWhenSet : ContentPage
@@ -88,7 +91,7 @@ namespace Prism.Forms.Tests.Mvvm
         {
             ResetViewModelLocationProvider();
 
-            PageMock view = new PageMock();
+            ViewModelLocatorPageMock view = new ViewModelLocatorPageMock();
             Assert.Null(view.BindingContext);
 
             string viewModel = "Test String";
@@ -97,6 +100,8 @@ namespace Prism.Forms.Tests.Mvvm
             ViewModelLocator.SetAutowireViewModel(view, true);
             Assert.NotNull(view.BindingContext);
             Assert.Equal(viewModel, view.BindingContext);
+
+            ResetViewModelLocationProvider();
         }
 
         private static void ResetViewModelLocationProvider()
