+++ Source/Xamarin/Prism.Forms.Tests/Prism.Forms.Tests.csproj
@@ -10,6 +10,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <None Include="Navigation\PageNavigationServiceFixture.NavigationPage.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
     <PackageReference Include="xunit" Version="2.3.0-rc1-build3809" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.0-rc1-build3809" />
